requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib
bcrypt
email-validator